feat: Track applied renames in workspace and enable mapping export
The workspace now keeps track of mappings applied to it in an aggregated
map conforming to the internal ASM format. This aggregated map can then
be used to provide a basic export facility for mappings in that format
(the Simple format).

The mapping in the workspace only stores the end result of all applied
mappings. That means that rename operations "a -> b -> c" are folded
together and stored as "a -> c". To enable this, AsmMappingUtils
contains code to apply new mapping rules to existing mappings by just
operating on the mapping Strings itself rather than a real class graph.

While it would be nice to be able to export to all the supported import
formats, this is a start and should be possible to expand on later.

This is a partial implementation of #281