Break out EXPR -> Bytecode into separate transformer, create new EXPR -> AST[] transformer
This feature will be required if we want to be able to analyze the AST since we can't inspect the behavior of an EXPR's contents. This transformer allows us to break it into a series of AST elements.

Additionally, this would be a nice util in the UI to have. User can type in Java code and see live bytecode results.