Override path node equals impls for slower types
We dont need to check for bundle equality if the references are the same. That idea expands to the workspace. Then for classes we can just do a name check and parent check instead of ensuring the exact class models are the same.

This drastically speeds up tree building when loading larger inputs.