Move throwable marking to the background
The main goal is to ensure the UI visibly has the exception icons. Before we did this in a blocking fashion. However, because those icons are loaded dynamically its very likely that we can do this in the background and have it complete before a normal user interacts with the UI in a way that would present throwable class icons.