Fix regression in 529061bc7 breaking ClassInfo equality checks
Because the info-importer is can be configured to skip method metadata (which it will by default) this broke the tests where the other model was using the metadata.

Realistically if two JVM classes have the same byte-array then they are equal, and we dont need to do any model checking. Only if the array is different should we fall back to model checks.