Fix phantom generation throwing illegal state when multiple keys overlap
This generally seems to occur when the javac service asks to build phantoms with classes that are in the primary resource of a workspace, and also present at runtime. Thus the "find class by name" yields two instances.

The basic merge using first should suffice since this should never occur for cases where the duplicate class is a target for editing. IE, users won't realistically be editing Recaf with Recaf, or core JDK classes.