Ensure transformers that can lead to blocks being marked as dead-code remove try-catches in those ranges
See the test case for an example. The opaque predicate is replaced with a goto during transformation. That makes the code for the try-catches dead, so when we remove the dead code we need to remove the try-catches because their label references are no longer in the method's instructions list.