Replace transformer collections of ClassPathNode with identity based collections over hash based collections
This massively improves performance due to the expensive nature of path hashing.

The creation of these collections is such that no duplicate paths should ever be attempted to go into the same collection.
Additionally the user is not expected to do path based lookups in these maps. They are for primarily intended for iteration/inspection only.